app/login-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx>=0.24.0,<0.25.0
python-dotenv==1.0.0
authlib==1.2.1
itsdangerous==2.1.2
supabase==1.2.0
postgrest>=0.10.8,<0.12.0 